Prepare a utils module for all Python utilities
