+++ python/rascaline/utils/__init__.py
@@ -1,0 +1,15 @@
+import os
+
+import equistore.core
+
+
+# path that can be used with cmake to access the rascaline library and headers
+_HERE = os.path.realpath(os.path.dirname(__file__))
+
+_rascaline_cmake_prefix = os.path.realpath(os.path.join(_HERE, "..", "lib", "cmake"))
+cmake_prefix_path = (
+    f"{_rascaline_cmake_prefix};{equistore.core.utils.cmake_prefix_path}"
+)
+"""
+Path containing the CMake configuration files for the underlying C library
+"""
